fix(core): make two over-promising docs match the code, and stop the lints
Clippy -D warnings flagged three unread fields and two useless casts. Two of the fields were a docs bug, not a style bug: API.md listed "session_id" in the /v1/ingest body (ingest must NOT buffer into the session, or the same transcript is extracted twice at flush) and showed "label" on the lock endpoint (ids are content-addressed from labels, so a rename would orphan the original node and leave it locked). Both structs now state what the field is for and why it is unread; API.md says the same thing to the outside; the sync handle documents that sync returns 501. The two casts follow rustc s own suggestion (confidence.clamp(0.0,1.0) with no cast; f32::from(impact.max(1))). report-cargo.sh also stopped being mute about assertion failures: it grepped only rustc diagnostics, so a run with a green compile and red tests came back as no diagnostic lines found. It now matches panic/assertion/failure-list lines and prints the failing test names; checked against a synthetic log.

+++ cortex-core/src/types/mod.rs
@@ -302,7 +302,7 @@ impl MemoryNode {
     /// Applies an extracted fact's salience to this node.
     pub fn reinforce_with_impact(&mut self, impact: u8, confidence: f32) {
         self.impact = self.impact.max(impact.clamp(1, 10));
-        let boost = 1.0 + confidence.clamp(0.0, 1.0) as f32;
+        let boost = 1.0 + confidence.clamp(0.0, 1.0);
         self.stability = (self.stability + boost).clamp(1.0, 100.0);
         self.mark_accessed();
     }
@@ -342,7 +342,7 @@ impl MemoryNode {
         let lock_bonus = if self.locked { 4.0 } else { 1.0 };
         let mesh_bonus = if is_global_mesh(&self.owner_uri) { 1.25 } else { 1.0 };
         (self.retention_probability().max(0.01)
-            * (self.stability.max(1.0) * f32::from(self.impact.max(1) as f32)).ln_1p()
+            * (self.stability.max(1.0) * f32::from(self.impact.max(1))).ln_1p()
             * lock_bonus
             * mesh_bonus)
             + (self.access_count as f32).ln_1p()
